Handles change of active tab while popup open
Learned that the activeTab permission only applies for the tab where the
popup was first opened. As such, the popup is closed whenever the active
tab changes. Could choose to use the "tabs" permission to allow the
popup to be kept open across tabs (would still have to handle change of
active tab when filtering is on).